+++ pyautocv/segmentation.py
@@ -3,6 +3,7 @@ import cv2
 import matplotlib.pyplot as plt
 from scipy import ndimage
 from skimage.io import imread_collection, imread
+from skimage.transform import resize
 from skimage import filters
 from os import pathsep
 import numpy as np
@@ -39,8 +40,10 @@ class Segmentation(object):
         self.image_suffix = image_suffix
         self.color_mode = color_mode
 
-    def read_images(self):
+    def read_images(self, other_directory=None):
         """
+        :param other_directory: If images exist in another folder/sub folder, please provide it here. Leave blank if
+        mixed file formats(jpg and png) exist in the same folder
 
         :return: Returns a multidimensional array containing arrays that represent images in a directory
 
@@ -52,7 +55,9 @@ class Segmentation(object):
         else:
             if self.image_suffix not in ["png", "jpg"]:
                 raise ValueError("Only tif, png, and jpg are currently supported")
-            return list(imread_collection(self.directory + "/*.jpg" + pathsep + "/*.png"))
+            if other_directory is None:
+                other_directory = self.directory
+            return list(imread_collection(self.directory + "/*.jpg" + pathsep + other_directory + "/*.png"))
 
     def smooth(self, mask="box", kernel_shape=(5, 5), **kwargs):
         """
@@ -157,12 +162,12 @@ def show_images(original_images=None, processed_images=None, cmap="gray", number
     :param cmap: Color cmap from matplotlib. Defaults to gray
     :param number: optional Number of images to show
     """
-    # need to figure out how any works in python
+
     if original_images is None or processed_images is None:
         raise ValueError("Both original and processed image lists are required.")
     if number is not None:
-        original_images = original_images[:number]
-        processed_images = processed_images[:number]
+        original_images = reshape_images(original_images[:number])
+        processed_images = reshape_images(processed_images[:number])
 
     image_list = list(chain(*zip(original_images, processed_images)))
 
@@ -175,3 +180,27 @@ def show_images(original_images=None, processed_images=None, cmap="gray", number
     for ind, image in enumerate(image_list):
         axes.ravel()[ind].imshow(image_list[ind], cmap=cmap)
         axes.ravel()[ind].set_axis_off()
+
+
+def resize_images(image_list, target_size):
+    """
+
+    :param image_list: A list of images or image that needs to be resized
+    :param target_size: New target image size
+    :return: Image or list of images that have been resized.
+
+
+    """
+
+    return [resize(x, target_size) for x in image_list]
+
+
+def reshape_images(image_list):
+    """
+
+    :param image_list: A list of images to reshape for plotting
+    :return: Images that can be plotted with show_images
+
+    """
+    final_list = [img[:, :, 0] if img.shape == 3 else img for img in image_list]
+    return final_list