+++ examples/example.py
@@ -3,7 +3,8 @@ import os
 os.chdir("pyautocv")
 # gray images
 images_list = Segmentation("images/dic", image_suffix="tif", color_mode="gray")
-show_images(images_list.read_images(), images_list.read_images(), number=4)
+
+
 # smooth images
 show_images(images_list.read_images(), images_list.smooth())
 # use median