+++ pyautocv/__init__.py
@@ -10,7 +10,7 @@ import os
 import sys
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 __author__="Nelson Gonzabato"
-__version__="0.2.0".encode("ascii","ignore").decode('ascii')
+__version__="0.3.0".encode("ascii","ignore").decode('ascii')
 __all__=["segmentation"]