+++ changelog.md
@@ -2,6 +2,8 @@
 
 **pyautocv 0.2.2**
 
+- `plot_hist` is a new function that allows one to plot histograms of images. 
+
 - `stack_images` is a new function that allows one to stack images vertically and horizontally.
 
 - `show_images` now has a titles argument to add titles to plots.