+++ changelog.md
@@ -1,5 +1,6 @@
 # pyautocv's change log 
 
+**Release 0.3.0**
 **pyautocv 0.2.2**
 
 - Extended tests and coverage